[Refactoring] 포스트 리스트 설계 구조 변경 및 pageNation 통일된 관리
