--- src/components/postList.module.scss
@@ -1,39 +1,0 @@
-@use 'src/styles/colors';
-
-.container {
-  width: 100%;
-  padding: 4em 2em;
-  background-color: colors.$BLACK;
-
-  @media only screen and (max-width: 820px) {
-    padding: 2em 1em;
-  }
-
-  .postList {
-    width: 90%;
-    max-width: 1050px;
-    margin: 0 auto;
-  }
-
-  h1 {
-    .title {
-      margin-right: 0.7rem;
-    }
-
-    .titleCount {
-      display: inline-block;
-      padding-right: 5px;
-      padding-left: 5px;
-      line-height: 1;
-      color: colors.$PUPLE;
-      background-color: colors.$WHITE;
-      box-shadow: 0 0 5px 0 rgb(0 0 0 / 5%);
-      transition: 0.2s ease-out;
-
-      &:hover {
-        color: colors.$WHITE;
-        background-color: colors.$PUPLE;
-      }
-    }
-  }
-}
